Addressed comments, added example for high-level client.
